+++ device/pin.c
@@ -137,7 +137,7 @@ static ssize_t mr_pin_write(struct mr_dev *dev, int off, const void *buf, size_t
     return wr_size;
 }
 
-static ssize_t mr_pin_ioctl(struct mr_dev *dev, int off, int cmd, void *args)
+static int mr_pin_ioctl(struct mr_dev *dev, int off, int cmd, void *args)
 {
     struct mr_pin *pin = (struct mr_pin *)dev;
