+++ device/spi.c
@@ -498,7 +498,7 @@ static ssize_t mr_spi_dev_write(struct mr_dev *dev, int off, const void *buf, si
     return ret;
 }
 
-static ssize_t mr_spi_dev_ioctl(struct mr_dev *dev, int off, int cmd, void *args)
+static int mr_spi_dev_ioctl(struct mr_dev *dev, int off, int cmd, void *args)
 {
     struct mr_spi_dev *spi_dev = (struct mr_spi_dev *)dev;
